celebal week 8/requirements.txt
streamlit==1.34.0
pandas==2.2.2
scikit-learn==1.7.0
sentence-transformers==2.2.2
# faiss-cpu @ https://dl.fbaipublicfiles.com/faiss/python/wheels/cpu/torch2.2/faiss_cpu-1.7.4-cp312-cp312-win_amd64.whl
torch==2.2.2
SpeechRecognition==3.10.1
gTTS==2.5.1
matplotlib==3.8.4
fpdf==1.7.2
langdetect==1.0.9
Pillow==10.3.0
sumy==0.11.0  
